Introduce PassengerAgent file-read-helper
A helper app that reads from an arbitrary file.
Its main reason for existance is to allow root processes (such as the Core)
to read from arbitrary files in a way that's safe from symlink and other
kinds of attacks. See the documentation for safeReadFile()
to learn more about the different types of attacks.

file-read-helper is used when the caller cannot use safeReadFile(),
e.g. when the following two conditions hold at the same time:

 1. The caller does not have control over the safety of the parent
    directories leading to the file.
 2. The caller cannot choose not to disclose the contents of the file.

file-read-helper MUST be used in combination with exec-helper in order
to lower its privilege, otherwise no protection is provided.